\name{caretAddOn-package}
\alias{caretAddOn-package}
\docType{package}
\title{Additional functionalities for the caret package}
\description{Additional functionalities for the caret package, including new models, new summaries, and new general-purpose functions.}
\details{
  \tabular{ll}{
  Package: \tab caretAddOn\cr
  Type: \tab Package\cr
  Version: \tab 0.1.1\cr
  Date: \tab 2022-05-13\cr
  License: \tab GPL-2
  }
New models:
\itemize{
  \item{\link{loglm_mean}, }{}
  \item{\link{loglm_median}, }{}
  \item{\link{svm_radial}, }{}
  }
New custom summaries:
\itemize{
  \item{\link{customSummaryClass}, }{}
  \item{\link{customSummaryReg}, }{}
  }
New general-purpose functions:
\itemize{
  \item{\link{addTerms}, }{}
  \item{\link{stepCV}, }{}
  \item{\link{bestTune}, }{}
  \item{\link{enetCoef}, }{}
  \item{\link{trainPlot}, }{}
  \item{\link{importancePlot}, }{}
  }
}
\author{Alessandro Magrini <alessandro.magrini@unifi.it>}
%\references{}
